The end..?
This is as much as I can do, there was a lot I wish I could've done but didn't do! I just want to submit this assessment thanks for everyone for trying even if it was minimal or large, I take this a huge learning improvement from here on out. I might work on it from time to time. Good foundations to make something great